V2.1.2 (#15)
* add 添加跨域，现在也可以用于前端

* 前端页面复制兼容老版本浏览器

* 修改 README.md

* 默认环境移除h2-console
（若需要调试数据库请使用dev环境）

* 添加flux

* 新增： YiMo后端支持stream返回

* 新增： YiMo前端支持stream输出

* 修复： 根据groupId查询配置列表会模糊匹配

* 前端交互优化

* 前端分组管理添加滚动分页

* 前端查询api为空时，隐藏分页组件

* 前端动画抖动修复

* 前端优化

+++ src/main/java/top/paidaxin/dao/entity/ApiConfig.java
@@ -54,7 +54,7 @@ public class ApiConfig extends ApiGroup implements Serializable {
     /**
      * 响应延迟(毫秒)
      */
-    private Integer delay = 0;
+    private Long delay = 0L;
 
     /**
      * 是否启用
